+++ subset.js
@@ -43,41 +43,30 @@ $.comparing = function () {
   };
 };
 
-// default equality, like `operators` eq2
-var eq2 = function (x, y) {
-  return x === y;
+// helper for set functions
+$.indexOfBy = function (eq, xs, x) {
+  for (var i = 0, len = xs.length; i < len; i += 1) {
+    if (eq(xs[i], x)) {
+      return i;
+    }
+  }
+  return -1;
 };
+
+// default equality, like `operators` eq2
+var eq2 = (x, y) => x === y;
 
 // ---------------------------------------------
 // Operations
 // ---------------------------------------------
 
 // max/min + generalized
-$.maximum = function (xs) {
-  return Math.max.apply(Math, xs);
-};
+$.maximum = (xs) => Math.max.apply(Math, xs);
+$.minimum = (xs) => Math.min.apply(Math, xs);
 
-$.minimum = function (xs) {
-  return Math.min.apply(Math, xs);
-};
-
-$.maximumBy = function (cmp, xs) {
-  for (var i = 1, max = xs[0], len = xs.length; i < len; i += 1) {
-    if (cmp(xs[i], max) > 0) {
-      max = xs[i];
-    }
-  }
-  return max;
-};
-
-$.minimumBy = function (cmp, xs) {
-  for (var i = 1, min = xs[0], len = xs.length; i < len; i += 1) {
-    if (cmp(xs[i], min) < 0) {
-      min = xs[i];
-    }
-  }
-  return min;
-};
+// TODO: dont crash on zero? haskell does this too..
+$.maximumBy = (cmp, xs) => xs.reduce((max, x) => cmp(x, max) > 0 ? x : max, xs[0]);
+$.minimumBy = (cmp, xs) => xs.reduce((min, x) => cmp(x, min) < 0 ? x : min, xs[0]);
 
 // Modifying Array operations
 $.insertBy = function (cmp, xs, x) {
@@ -90,82 +79,38 @@ $.insertBy = function (cmp, xs, x) {
   xs.push(x);
   return xs;
 };
-
-$.insert = function (xs, x) {
-  return $.insertBy($.compare(), xs, x);
-};
+$.insert = (xs, x) => $.insertBy($.compare(), xs, x);
 
 $.deleteBy = function (eq, xs, x) {
-  for (var i = 0, len = xs.length; i < len; i += 1) {
-    if (eq(xs[i], x)) {
-      xs.splice(i, 1);
-      return xs;
-    }
-  }
-  return xs;
-};
-
-$.delete = function (xs, x) {
-  var idx = xs.indexOf(x);
+  var idx = $.indexOfBy(eq, xs, x);
   if (idx >= 0) {
     xs.splice(idx, 1);
   }
   return xs;
 };
+$.delete = (xs, x) => $.deleteBy(eq2, xs, x);
 
 // "Set" operations
 $.intersectBy = function (eq, xs, ys) {
-  var result = []
-    , xLen = xs.length
-    , yLen = ys.length;
-
-  if (xLen && yLen) {
-    for (var i = 0; i < xLen; i += 1) {
-      var x = xs[i];
-      for (var j = 0; j < yLen; j += 1) {
-        if (eq(x, ys[j])) {
-          result.push(x);
-          break;
-        }
-      }
+  return xs.reduce((acc, x) => {
+    if ($.indexOfBy(eq, ys, x) >= 0) {
+      acc.push(x);
     }
-  }
-  return result;
+    return acc;
+  }, []);
 };
-
-$.intersect = function (xs, ys) {
-  return $.intersectBy(eq2, xs, ys);
-};
+$.intersect = (xs, ys) => $.intersectBy(eq2, xs, ys);
 
 $.nubBy = function (eq, xs) {
-  var result = [];
-  for (var i = 0, len = xs.length; i < len; i += 1) {
-    var keep = true;
-    for (var j = 0, resLen = result.length; j < resLen; j += 1) {
-      if (eq(result[j], xs[i])) {
-        keep = false;
-        break;
-      }
+  return xs.reduce((acc, x) => {
+    if ($.indexOfBy(eq, acc, x) < 0) {
+      acc.push(x);
     }
-    if (keep) {
-      result.push(xs[i]);
-    }
-  }
-  return result;
+    return acc;
+  }, []);
 };
+$.nub = (xs) => $.nubBy(eq2, xs);
 
-// nub, build up a list of unique (w.r.t. equality)
-// elements by checking if current is not 'equal' to anything in the buildup
-// http://jsperf.com/nubnubbytest1 => indexOf clearly beats calling $.nubBy(eq2)
-$.nub = function (xs) {
-  var result = [];
-  for (var i = 0, len = xs.length; i < len; i += 1) {
-    if (result.indexOf(xs[i]) < 0) {
-      result.push(xs[i]);
-    }
-  }
-  return result;
-};
 
 $.groupBy = function (eq, xs) {
   var result = []
@@ -179,38 +124,32 @@ $.groupBy = function (eq, xs) {
   }
   return result;
 };
-
-$.group = function (xs) {
-  return $.groupBy(eq2, xs);
+$.groupBy = function (eq, xs) {
+  var result = []
+    , j, sub;
+  for (var i = 0, len = xs.length; i < len; i = j) {
+    sub = [xs[i]];
+    for (j = i + 1; j < len && eq(xs[i], xs[j]); j += 1) {
+      sub.push(xs[j]);
+    }
+    result.push(sub);
+  }
+  return result;
 };
+$.group = (xs) => $.groupBy(eq2, xs);
 
 $.unionBy = function (eq, xs, ys) {
   return xs.concat(xs.reduce($.deleteBy.bind(null, eq), $.nubBy(eq, ys)));
 };
+$.union = (xs, ys) => xs.concat(xs.reduce($.delete, $.nub(ys)));
 
-$.union = function (xs, ys) {
-  return xs.concat(xs.reduce($.delete, $.nub(ys)));
-};
+$.differenceBy = (eq, xs, ys) => ys.reduce($.deleteBy.bind(null, eq), xs.slice());
+$.difference = (xs, ys) => ys.reduce($.delete, xs.slice());
 
-$.differenceBy = function (eq, xs, ys) {
-  return ys.reduce($.deleteBy.bind(null, eq), xs.slice()); // reduce a copy
-};
-
-$.difference = function (xs, ys) {
-  return ys.reduce($.delete, xs.slice());
-};
-
-$.isSubsetOf = function (xs, ys, proper) {
-  var parent = ys.slice();
-  for (var i = 0; i < xs.length; i += 1) {
-    var x = xs[i]
-      , idx = parent.indexOf(x);
-    if (idx < 0) {
-      return false;
-    }
-    parent.splice(idx, 1);
-  }
-  return (proper) ? (parent.length > 0) : true;
+// These now assume no duplicates
+$.isSubsetOf = (xs, ys) => xs.every((x) => ys.indexOf(x) >= 0);
+$.isProperSubsetOf = function (xs, ys) {
+  return $.isSubsetOf(xs, ys) && ys.some((y) => xs.indexOf(y) < 0);
 };
 
 // end - export